+++ mysql_db_admin.py
@@ -572,6 +572,7 @@ def status(server, args_array, **kwargs):
             gen_libs.print_data(jdata)
 
     else:
+        ofile = kwargs.get("ofile", None)
 
         if not args_array.get("-z", False):
             print("\nDatabase Status Check for Server: %s" % (server.name))
@@ -584,6 +585,15 @@ def status(server, args_array, **kwargs):
             gen_libs.prt_msg("Max Connections", server.max_conn, 1)
             gen_libs.prt_msg("Current Connections", server.cur_conn, 1)
             gen_libs.prt_msg("Percent Used", server.prct_conn, 1)
+
+        if ofile:
+            pdata = ""
+
+            for key, value in outdata.items():
+                pdata += "{}: {}".format(key, value) + "\n"
+
+            gen_libs.write_file(ofile, mode, pdata)
+            
 
 
 def run_program(args_array, func_dict, **kwargs):